# backend/src/test/resources/contracts/openapi.yaml
openapi: 3.0.3
info:
  title: 교수업적평가 시스템 공통기능 API
  version: 1.0.0
  description: 1차 시스템 관리 영역의 HTTP API 계약이다.
servers:
- url: /
  description: 브라우저 상대경로 API 기준
paths:
  /api/auth/login:
    post:
      operationId: login
      summary: 내부 관리자 로그인
      x-related-requirements:
      - REQ-030
      x-business-rules:
      - 시드 관리자 계정으로 로그인하면 R09 권한으로 1차 목표 9개 메뉴를 테스트할 수 있어야 한다
      x-side-effects:
      - 인증에 성공한 시드 관리자에 대해 활성 session을 생성한다
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200': &id001
          description: 성공 응답
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
        '400':
          description: 로그인 입력 오류
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '401':
          description: 로그인 실패
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
      x-required-tests:
      - 'happy: 유효한 admin/admin 자격 증명으로 로그인 요청 -> 성공 ApiResponse envelope와 인증 세션 cookie 반환'
      - 'auth: 존재하지 않는 사용자 자격 증명으로 로그인 요청 -> 401과 인증 세션 cookie 미발급'
      - 'validation: username 또는 password 누락 로그인 요청 -> 400과 ApiError fieldErrors 반환'
      - 'business: INACTIVE 사용자 자격 증명으로 로그인 요청 -> 401과 인증 세션 cookie 미발급'
      - 'side-effect: 성공 로그인 요청 -> session이 ACTIVE 상태로 생성되고 getCurrentUser로 현재 사용자 조회 가능'
  /api/auth/logout:
    post:
      operationId: logout
      summary: 세션 로그아웃
      security:
      - sessionCookie: []
      x-related-requirements:
      - REQ-030
      responses:
        '200': *id001
        '401':
          description: 인증 필요
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '403':
          description: 권한 없음
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '400':
          description: 입력 오류
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
      x-roles:
      - R09
      x-business-rules:
      - 인증된 활성 세션만 로그아웃할 수 있다
      x-side-effects:
      - 현재 session 상태를 INVALIDATED로 변경한다
      x-required-tests:
      - 'happy: 활성 세션으로 로그아웃 요청 -> session.status가 INVALIDATED로 변경됨'
      - 'auth: 세션 cookie 없이 로그아웃 요청 -> 401 반환'
      - 'validation: 형식이 아닌 세션 cookie로 로그아웃 요청 -> ApiError가 아닌 성공 응답 envelope 반환'
      - 'business: 이미 INVALIDATED인 세션으로 로그아웃 요청 -> 400과 session.status 미변경'
      - 'side-effect: 성공 요청 -> 현재 session 행의 updated_at이 로그아웃 시각으로 갱신됨'
      x-state-transitions:
      - ACTIVE -> INVALIDATED
  /api/auth/me:
    get:
      operationId: getCurrentUser
      summary: 현재 사용자 조회
      security:
      - sessionCookie: []
      x-related-requirements:
      - REQ-030
      responses:
        '200': *id001
        '401':
          description: 인증 필요
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '403':
          description: 권한 없음
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '400':
          description: 입력 오류
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
      x-roles:
      - R01
      - R02
      - R03
      - R04
      - R05
      - R06
      - R07
      - R08
      - R09
...
  /api/code-groups:
    get:
      operationId: listCodeGroups
      summary: 코드그룹 조회
      security:
      - sessionCookie: []
      x-related-requirements:
      - REQ-022
      responses:
        '200': *id001
        '401':
          description: 인증 필요
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '403':
          description: 권한 없음
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '400':
          description: 입력 오류
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
      parameters:
      - name: page
        in: query
        required: false
        schema:
          type: integer
          minimum: 0
      - name: size
        in: query
        required: false
        schema:
          type: integer
          minimum: 1
      x-roles:
      - R09
    post:
      operationId: createCodeGroup
      summary: 코드그룹 등록
      security:
      - sessionCookie: []
      x-related-requirements:
      - REQ-023
      - REQ-028
      - REQ-029
      responses:
        '200': *id001
        '401':
          description: 인증 필요
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '403':
          description: 권한 없음
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '400':
          description: 입력 오류
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CodeGroupRequest'
      x-roles:
      - R09
      x-business-rules:
      - group_id는 필수이며 중복될 수 없다
      x-side-effects:
      - code_group을 ACTIVE로 생성한다
      - change_history에 코드그룹 등록을 기록한다
      x-required-tests:
      - 'happy: 유효한 R09 요청 -> code_group의 group_id 변경을 후속 GET으로 확인'
      - 'auth: R09 이외의 인증 사용자 요청 -> 403과 code_group 미변경'
      - 'validation: 필수 group_id 누락 -> ApiError fieldErrors에 group_id 오류'
      - 'business: group_id 중복 금지 위반 요청 -> 400과 대상 상태 미변경'
      - 'side-effect: 성공 요청 -> change_history에 코드그룹 변경 이력 존재'
  /api/code-groups/{groupId}:
    put:
      operationId: updateCodeGroup
      summary: 코드그룹 변경
      security:
      - sessionCookie: []
      x-related-requirements:
      - REQ-023
      - REQ-028
      - REQ-029
      responses:
        '200': *id001
        '401':
          description: 인증 필요
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '403':
          description: 권한 없음
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '400':
          description: 입력 오류
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
      parameters:
      - name: groupId
        in: path
        required: true
        schema: *id002
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CodeGroupRequest'
      x-roles:
      - R09
      x-business-rules:
      - path의 group_id와 요청 대상은 일치해야 한다
      x-side-effects:
      - code_group 필드를 수정한다
      - change_history에 코드그룹 변경을 기록한다
      x-required-tests:
      - 'happy: 유효한 R09 요청 -> code_group의 group_name 변경을 후속 GET으로 확인'
      - 'auth: R09 이외의 인증 사용자 요청 -> 403과 code_group 미변경'
      - 'validation: 필수 group_name 누락 -> ApiError fieldErrors에 group_name 오류'
      - 'business: 경로 group_id 일치 위반 요청 -> 400과 대상 상태 미변경'
      - 'side-effect: 성공 요청 -> change_history에 코드그룹 변경 이력 존재'
...
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: SESSION
      description: HttpOnly SameSite=Lax 세션 cookie
  schemas:
...
    ApiResponse:
      type: object
      required:
      - success
      - meta
      properties:
        success:
          type: boolean
          enum:
          - true
        data: {}
        meta:
          type: object
          additionalProperties: true
    ApiError:
      type: object
      required:
      - success
      - meta
      properties:
        success:
          type: boolean
          enum:
          - false
        error:
          type: object
          properties:
            code:
              type: string
            message:
              type: string
            fieldErrors:
              type: array
              items:
                type: object
        meta:
          type: object
          additionalProperties: true
    LoginRequest:
      type: object
      required:
      - username
      - password
      properties:
        username:
          type: string
        password:
          type: string
...
    CodeGroupRequest:
      type: object
      required:
      - groupId
      - groupName
      - reason
      properties:
        groupId:
          type: string
        groupName:
          type: string
        description:
          type: string
        managingDepartment:
          type: string
        reason:
          type: string
...